+++ lib/enums/animation_type.dart
@@ -1,0 +1,9 @@
+enum AnimateType {
+  fade,
+  scale,
+  slideUp,
+  slideDown,
+  slideLeft,
+  slideRight,
+  rotate,
+}